added-components
splited to components
